requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
numpy==2.1.1
openai==1.47.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pip>=22.3.1
pydantic-core==2.23.4
pydantic==2.9.2
pytesseract==0.3.13
screeninfo==0.8.1
setuptools==65.5.1
sniffio==1.3.1
tqdm==4.66.5
typing-extensions==4.12.2
wheel==0.38.4
